refactor(app): move the sample-rate queries into their own file
AppAudioCapture.swift sat at 591 lines against a 600-line lint cap, with
no room for the delivered-rate measurement that follows. The five static
rate queries are the natural block to lift out: they are pure CoreAudio
property reads plus one call into SampleRateQuery, and nothing else in
the file touches their internals.

Pure extraction. The main file's diff is 114 deletions and no additions.
The only change to the moved code is that `private static` becomes
`static`, because the callers stay behind in AppAudioCapture.swift and
`private` at type scope does not reach across files.
